refactor composable
