+++ composables/renderers/index.ts
@@ -1,0 +1,2 @@
+export { createMermaidRenderer } from './mermaid';
+export { createMathBlockRenderer, createMathInlineRenderer } from './katex';
